Data.Fits.Read

Documentation

readHDUs :: ByteString -> Either String [HeaderDataUnit]

Parse and read all HDUs in the input string

readPrimaryHDU :: ByteString -> Either String HeaderDataUnit

Parse and read only the Primary HDU from the input string

getKeyword :: Text -> (Value -> Maybe a) -> HeaderDataUnit -> Either String a

Look up a keyword and parse it into the expected format

getHDU :: String -> Int -> [HeaderDataUnit] -> Either String HeaderDataUnit

Get the HDU at an index and fail with a readable error

maybeError :: FitsError -> Maybe a -> Either String a

eitherFail :: MonadFail m => Either String a -> m a

data FitsError

Constructors

ParseError ParseErr

MissingKey Keyword

InvalidKey Keyword Value

MissingHDU String Int

InvalidData String