HeaderDataUnit

Description

Definitions for the data types needed to parse an HDU in a FITS block.

Documentation

type Bytes = Int

type Count = Int

hduRecordLength :: Bytes

hduMaxRecords :: Count

hduBlockSize :: Bytes

data StringType

Constructors

NullString

EmptyString

DataString

UndefinedString

data StringValue

Constructors

StringValue

Fields

stringType :: StringType

stringValue :: Maybe Text

data NumberType

Constructors

IntegerType
HDU ASCII encoded integer number

RealType
HDU ASCII encoded real number

ComplexType
HDU ASCII encoded complex number

data NumberModifier

Constructors

Positive
HDU positive number value

Negative
HDU negative number value

Zero
HDU numeric value is zero, could be positive or negative

data NumberValue

NumberValue contains an encoded numeric record from a data field. This data type still needs to be converted into more useful Haskell data types.

Constructors

NumberValue

Fields

numberType :: NumberType

realModifier :: NumberModifier

realPart :: Text

imaginaryModifier :: Maybe NumberModifier

imaginaryPart :: Maybe Text

exponentModifier :: Maybe NumberModifier

exponent :: Maybe Int

data SimpleFormat

Constructors

Conformant
Value of SIMPLE is T in the header

NonConformant
Value of SIMPLE is F in the header

data NAxisType

Constructors

ZeroAxes
No data follows the header

ManyAxes
There is one or more axes in the data following the header

data NAxisMetadata

Constructors

NAxisMetadata

Fields

naxisType :: NAxisType

Are there zero or more axes of data?

axesCount :: Int

How many axes are there in the dataset?

data Axis

Constructors

Axis

Fields

axisNumber :: Int

The axis number under consideration

axisElementCount :: Int

The number of elements in this axis

data BitPixFormat

Constructors

EightBitInt
BITPIX = 8; unsigned binary integer of 8 bits

SixteenBitInt
BITPIX = 16; two's complement binary integer of 16 bits

ThirtyTwoBitInt
BITPIX = 32; two's complement binary integer of 32 bits

SixtyFourBitInt
BITPIX = 64; two's complement binary integer of 64 bits

ThirtyTwoBitFloat
BITPIX = -32; IEEE single precision floating point of 32 bits

SixtyFourBitFloat
BITPIX = -64; IEEE double precision floating point of 64 bits