Data.Fits

Description

Definitions for the data types needed to parse an HDU in a FITS block.

Data payload functions

parsePix :: Int -> BitPixFormat -> ByteString -> IO [Pix]

This is the main low-level function which parses the data portion of an HDU. You need and element count, a format and a bytestring. The resulting list is produced in column-row major order as specified in the standard.

pixsUnwrapI :: BitPixFormat -> [Pix] -> [Int]

Remove the Pix wrapper for integer Pix lists.

pixsUnwrapD :: BitPixFormat -> [Pix] -> [Double]

Remove the Pix wrapper for floating point Pix lists.

Main data types

data HeaderDataUnit

The HeaderDataUnit is the full HDU. Both the header information is encoded alongside the data payload.

Constructors

HeaderDataUnit

Fields

_header :: Header

The heeader contains metadata about the payload

_dimensions :: Dimensions

This dimensions of the main data array

_extension :: Extension

Extensions may vary the data format

_mainData :: ByteString

The main data array

dimensions :: Lens' HeaderDataUnit Dimensions

header :: Lens' HeaderDataUnit Header

extension :: Lens' HeaderDataUnit Extension

mainData :: Lens' HeaderDataUnit ByteString

data Pix

Following BitPixFormat we have a tag for integer and floating point values. We box them up to ease parsing.

Constructors

PB Int

PI16 Int

PI32 Int

PI64 Int

PF Double

PD Double

Header Data Types

newtype Header

The header part of the HDU is vital carrying not only authorship metadata, but also specifying how to make sense of the binary payload that starts 2,880 bytes after the start of the HeaderData.

Constructors

Header

Fields

_keywords :: Map Keyword Value

keywords :: Lens' Header (Map Keyword Value)

data Extension

Constructors

Primary
Any header data unit can use the primary format. The first MUST be Primary. This is equivalent to having no extension

Image
An encoded image. PCOUNT and GCOUNT are required but irrelevant

BinTable
A Binary table. PCOUNT is the number of bytes that follow the data in the heap

Fields

pCount :: Int

heap :: ByteString

lookup :: Keyword -> Header -> Maybe Value

newtype Keyword

The Text wrapper for HDU the keyword data for lines of the form: KEYWORD=VALUE

Constructors

Keyword Text

data Value

Value datatype for discriminating valid FITS KEYWORD=VALUE types in an HDU.

Constructors

Integer Int

Float Float

String Text

Logic LogicalConstant

toInt :: Value -> Maybe Int

toFloat :: Value -> Maybe Float

toText :: Value -> Maybe Text

data LogicalConstant

Direct encoding of a Bool for parsing Value

Constructors

T

F

data Dimensions

When we load a header, we parse the BITPIX and NAXIS(N) keywords so we - can know how long the data array is

Constructors

Dimensions

Fields

_bitpix :: BitPixFormat

_axes :: Axes

axes :: Lens' Dimensions Axes

bitpix :: Lens' Dimensions BitPixFormat

newtype Comment

Constructors

Comment Text

data SimpleFormat

The standard defines two possible values for the SIMPLE keyword, T and F. The T refers to a Conformant format while F refers to a NonConformant format. At this time only the Conformant, T, format is supported.

Constructors

Conformant

NonConformant

data BitPixFormat

The BitPixFormat is the nitty gritty of how the Axis data is layed out in the file. The standard recognizes six formats: unsigned 8 bit integer, two's complement binary integers at 16, 32, and 64 bits along with 32 and 64 bit IEEE floating point formats.

Constructors

EightBitInt
BITPIX = 8; unsigned binary integer of 8 bits

SixteenBitInt
BITPIX = 16; two's complement binary integer of 16 bits

ThirtyTwoBitInt
BITPIX = 32; two's complement binary integer of 32 bits

SixtyFourBitInt
BITPIX = 64; two's complement binary integer of 64 bits

ThirtyTwoBitFloat
BITPIX = -32; IEEE single precision floating point of 32 bits

SixtyFourBitFloat
BITPIX = -64; IEEE double precision floating point of 64 bits

type Axes = [Int]

Axes represents the combination of NAXIS + NAXISn. The spec supports up to 999 axes

Utility

isBitPixInt :: BitPixFormat -> Bool

This utility functions quickly lets you know if you are dealing with integer data.

isBitPixFloat :: BitPixFormat -> Bool

This utility functions quickly lets you know if you are dealing with floating point data.

bitPixToWordSize :: BitPixFormat -> Int

This utility function can be used to get the word count for data in an HDU.

bitPixToByteSize :: BitPixFormat -> Int

This utility function can be used to get the size in bytes of the - format.

pixDimsByCol :: Axes -> [Int]

pixDimsByRow :: Axes -> [Int]

Constants

hduRecordLength :: Int

A single record in the HDU is an eighty byte word.

hduMaxRecords :: Int

The maximum amount of eighty byte records is thirty-six per the standard.

hduBlockSize :: Int

The size of an HDU block is fixed at thirty-six eighty byte words. In other words 2,880 bytes. These blocks are padded with zeros to this boundary.